% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R, R/output_multiloc.R
\name{add_age}
\alias{add_age}
\title{Keep age groups}
\usage{
add_age(x)

add_age(x)
}
\arguments{
\item{x}{age-disaggregated odin output matrix or array
- 2D: time x age
- 3D: time x age x location}
}
\value{
age-disaggregated output matrix

long matrix with columns:
  - age_index, value              (single location)
  - age_index, location_index, value (multiple locations)
}
\description{
Keep age groups

Keep age groups (and locations, if present)
}
