% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R, R/output_multiloc.R
\name{collapse_age}
\alias{collapse_age}
\title{Collapse age groups}
\usage{
collapse_age(x)

collapse_age(x)
}
\arguments{
\item{x}{age-disaggregated odin output matrix or array
- 2D: time x age
- 3D: time x age x location}
}
\value{
age-aggregated output matrix

age-aggregated output:
  - 2D input: matrix with columns "value"
  - 3D input: matrix with columns "location_index", "value"
}
\description{
Collapse age groups

Collapse age groups (preserving locations if present)
}
