% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{countries}
\alias{countries}
\title{Vaccine parameters}
\format{
An object of class \code{character} of length 2.
}
\usage{
countries
}
\description{
Vaccine parameters
}
\details{
All durations are in days.
}
\keyword{datasets}
