% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{format_internal}
\alias{format_internal}
\title{Internal helper for formatting single-location output}
\usage{
format_internal(x, compartments, reduce_age, index, time, replicate)
}
\arguments{
\item{x}{nimue_simulation object returned by \code{run()}.}

\item{compartments}{Vector of compartment names present in the model output
  (e.g. \code{c("S", "R")}) or sub-compartment names (e.g.
  \code{c("S", "E1", "E2")}).}

\item{reduce_age}{Logical; if \code{TRUE} (default) collapse the age
  dimension, otherwise retain age-specific counts.}

\item{index}{odin output index}

\item{time}{Time vector from the odin model output}

\item{replicate}{Output replicate number}
}
\description{
Internal helper for formatting single-location output
}
