% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_multiloc.R
\name{format_internal_multiloc}
\alias{format_internal_multiloc}
\title{Internals of Format vaccine model output as data.frame}
\usage{
format_internal_multiloc(x, compartments, reduce_age, index, time, replicate)
}
\arguments{
\item{x}{squire_simulation object}

\item{compartments}{Vector of compartment names, e.g. \code{c("S", "R")}, or sub-compartment names, e.g. \code{c("S", "E1", "E2")}}

\item{reduce_age}{Collapse age-dimension, calculating the total in the
compartment.}

\item{index}{odin output index}

\item{time}{time vector}

\item{replicate}{output replicate number}
}
\description{
Internals of Format vaccine model output as data.frame
}
