% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrices.R
\name{matrix_check}
\alias{matrix_check}
\title{Check dimensions of inputs}
\usage{
matrix_check(population, contact_matrix_set)
}
\value{
Null if checks pass
}
\description{
Check dimensions of inputs
}
