% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Vaccine parameters}
\usage{
parameters(
  countries = NULL,
  R0 = 3,
  tt_R0 = 0,
  beta_set = NULL,
  time_period = 365,
  seeding_cases,
  seeding_age_order = NULL,
  init = NULL,
  prob_hosp = probs$prob_hosp,
  prob_death_hosp = probs$prob_death_hosp,
  p_dist = probs$p_dist,
  rel_infectiousness = probs$rel_infectiousness,
  rel_infectiousness_vaccinated = probs$rel_infectiousness_vaccinated,
  dur_E,
  dur_IMild,
  dur_ICase,
  dur_IHosp,
  dur_V,
  vaccine_efficacy_infection,
  tt_vaccine_efficacy_infection,
  vaccine_efficacy_disease,
  tt_vaccine_efficacy_disease,
  max_vaccine,
  tt_vaccine,
  dur_vaccine_delay,
  vaccine_coverage_mat
)
}
\arguments{
\item{R0}{Basic Reproduction Number. Default = 3}

\item{tt_R0}{Change time points for R0. Default = 0}

\item{beta_set}{Alternative parameterisation via beta rather than R0.
Default = NULL, which causes beta to be estimated from R0}

\item{time_period}{Length of simulation. Default = 365}

\item{seeding_cases}{Initial number of cases seeding the epidemic}

\item{seeding_age_order}{Vector specifying the order in which seeds are allocated to ages.
If NULL, seeds are distributed randomly within working ages. If specified, must be a vector
of length 17 specifying the order seeds are allocated, e.g. 1:17 will allocate first seed
to the youngest age group, then the second youngest and so on. Default = NULL}

\item{init}{Initial conditions for simulation provided. Allows overriding
if initial conditions start with an already infected population etc.
Default = NULL.}

\item{prob_hosp}{probability of hospitalisation by age.
Default = c(0.000744192, 0.000634166,0.001171109, 0.002394593, 0.005346437,
0.010289885, 0.016234604, 0.023349169, 0.028944623, 0.038607042,
0.057734879, 0.072422135, 0.101602458, 0.116979814, 0.146099064,
0.176634654 ,0.180000000)}

\item{p_dist}{Preferentiality of age group receiving treatment relative to
other age groups when demand exceeds healthcare capacity.}

\item{rel_infectiousness}{Relative infectiousness per age category relative
to maximum infectiousness category. Default = rep(1, 17)}

\item{rel_infectiousness_vaccinated}{Relative infectiousness per age
category  of vaccinated individuals relative to unvaccinated individuals.
Default = rep(1, 17), which is no impact of vaccination on onwards
transmissions}

\item{dur_E}{Mean duration of incubation period (days). Default = 4.6}

\item{dur_IMild}{Mean duration of mild infection (days). Default = 2.1}

\item{dur_ICase}{Mean duration from symptom onset to hospitil admission (days).
Default = 4.5}

\item{dur_V}{Mean duration of vaccine-derived immunity (days)}

\item{vaccine_efficacy_infection}{Efficacy of vaccine against infection.
This parameter must either be length 1 numeric (a single efficacy for all
age groups) or length 17 numeric vector (an efficacy for each age group).
An efficacy of 1 will reduce FOI by 100 percent, an efficacy of 0.2 will
reduce FOI by 20 percent etc.
To specify changes in vaccine efficacy over time, vaccine efficacies must
be provided as a list, with each list element being the efficacy at each
time point specified by \code{tt_vaccine_efficacy_infection}. These
efficacies must also be length 1 numeric (a single efficacy for all age
groups)  or length 17 numeric vector (an efficacy for each age group)}

\item{tt_vaccine_efficacy_infection}{Timing of changes in vaccine efficacy
against infection. Default = 0, which assumes fixed efficacy over time.
Must be the same length as the length of \code{vaccine_efficacy_infection}
when provided as a list. Time changing efficacies can occur in response to
changing vaccines being  given and dosing strategy changes.}

\item{vaccine_efficacy_disease}{Efficacy of vaccine against severe
(requiring hospitilisation) disease (by age). This parameter must either be
length 1 numeric (a single efficacy for all age groups) or length 17
numeric vector (an efficacy for each age group). An efficacy of 1 will
reduce the probability of hospitalisation by 100 percent, an efficacy of
0.2 will reduce the probability of hospitalisation by 20 percent etc.
To specify changes in vaccine efficacy over time, vaccine efficacies must
be provided as a list, with each list element being the efficacy at each
time point specified by \code{tt_vaccine_efficacy_disease}. These
efficacies must also be length 1 numeric (a single efficacy for all age
groups)  or length 17 numeric vector (an efficacy for each age group).}

\item{tt_vaccine_efficacy_disease}{Timing of changes in vaccine efficacy
against disease. Default = 0, which assumes fixed efficacy over time.
Must be the same length as the length of \code{vaccine_efficacy_disease}
when provided as a list. Time changing efficacies can occur in response to
changing vaccines being  given and dosing strategy changes.}

\item{max_vaccine}{The maximum number of individuals who can be vaccinated per day.}

\item{tt_vaccine}{Time change points for vaccine capacity (\code{max_vaccine}).}

\item{dur_vaccine_delay}{Mean duration of period from vaccination to vaccine protection.}

\item{vaccine_coverage_mat}{Vaccine coverage targets by age (columns) and priority (row)}
}
\description{
Vaccine parameters
}
\details{
All durations are in days.
}
