% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_multiloc.R
\name{format_multiloc}
\alias{format_multiloc}
\title{Format vaccine model output}
\usage{
format_multiloc(
  x,
  compartments = c("S", "E", "IMild", "ICase", "IHosp", "R", "D"),
  reduce_age = TRUE,
  date_0 = NULL,
  replicate = 1
)
}
\arguments{
\item{x}{nimue_simulation object}

\item{compartments}{Vector of compartment names, e.g. \code{c("S", "R")}, or sub-compartment names,
e.g. \code{c("S", "E1", "E2")}}

\item{reduce_age}{Collapse age-dimension, calculating the total in the
compartment (location is always retained if present).}

\item{date_0}{Date of time 0 (e.g. "2020-03-01"), if specified a date column will be added}

\item{replicate}{Which replicate is being formatted. Default = 1}
}
\value{
Formatted long data.frame
}
\description{
Take raw odin vaccine model output and formats in long format with the option to select
variables and summarise over age groups. Output variables are ordered as in argument ordering.
}
